+++ sql_scripts/create/reservation_extra.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS reservation_extra;
 CREATE TABLE reservation_extra(
-    id_reservation_extra INT AUTO_INCREMENT PRIMARY KEY,
-    id_reservation INT,
-    id_extra INT,
+    id_reservation_extra BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id_reservation BIGINT,
+    id_extra BIGINT,
     FOREIGN KEY (id_reservation) REFERENCES reservation(id_reservation),
     FOREIGN KEY (id_extra) REFERENCES extra(id_extra)
 );
